python/chat/requirements.txt
grpcio-tools==1.63.0
pymilvus==2.4.4
protobuf==3.20.0